fixed wx.Timer error
fixes #992

+++ MAVProxy/modules/lib/live_graph_ui.py
@@ -25,7 +25,7 @@ class GraphFrame(wx.Frame):
 
         self.redraw_timer = wx.Timer(self)
         self.Bind(wx.EVT_TIMER, self.on_redraw_timer, self.redraw_timer)
-        self.redraw_timer.Start(1000*self.state.tickresolution)
+        self.redraw_timer.Start(int(1000*self.state.tickresolution))
 
         self.last_yrange = (None, None)
